Pop share modal from header dropdown
