+++ components/form-builder/app/index.ts
@@ -9,5 +9,4 @@ export { Preview } from "./Preview";
 export { Publish } from "./Publish";
 export { Published } from "./Published";
 export { Edit } from "./edit";
-export { ShareButton } from "./share/ShareButton";
-export { ShareModal } from "./share/ShareModal";
+export { ShareModal } from "./ShareModal";
